Update Compose.js
Instead of global variables there's now a Compose class whose member functions are the functions that used to live at global scope in Compose.js

I had to prefix lots of variables with  this.  and I had to use bind to get the 'this' object passed to the doAnimation function.